[CMAKE] Find gapy instead of using `PATH`
The cmake script previously relied on `gapy` needing to be in the PATH. This makes the cmake script less self-contained and adds a magic ceremonial step required to configure the cmake build.

This PR therefore uses cmake's `find_program` to find the `gapy` script instead.
This has the following advantages:
* We can specify hints where cmake should try and find `gapy`. This way we can add the path from the monorepo layout as a convenient default
* If not found through hints, `find_program` will look in `PATH` preserving previous behavior
* The `gapy` executable can also be overwritten on the command line using `-DGAPY_SCRIPT_PATH=...`
* If not found, a nicer error message is omitted by cmake

Additionally, the `gapy` script is now executed using the python interpreter CMake was configured with, not the one that is currently active. This should better ensure that it is always using a venv with the requirements installed if desired.